+++ kpops/components/streams_bootstrap/producer/producer_app.py
@@ -81,9 +81,12 @@ class ProducerApp(StreamsBootstrap):
     @computed_field
     @cached_property
     def _cleaner(self) -> ProducerAppCleaner:
-        cleaner = ProducerAppCleaner(
-            **self.model_dump(by_alias=True, exclude={"_cleaner", "from_", "to"})
-        )
+        kwargs = {
+            name: getattr(self, name)
+            for name in self.model_fields_set
+            if name not in {"_cleaner", "from_", "to"}
+        }
+        cleaner = ProducerAppCleaner.model_validate(kwargs)
         cleaner.values.name_override = None
         return cleaner