+++ tests/utils/test_pydantic.py
@@ -2,7 +2,15 @@ from typing import Any
 
 import pytest
 
-from kpops.utils.pydantic import to_dash, to_dot, to_snake, to_str
+from kpops.components.common.kubernetes_model import SerializeAsOptional
+from kpops.utils.pydantic import (
+    SerializeAsOptionalModel,
+    exclude_by_value,
+    to_dash,
+    to_dot,
+    to_snake,
+    to_str,
+)
 
 
 @pytest.mark.parametrize(
@@ -64,3 +72,100 @@ def test_to_dot(input: str, expected: str):
 )
 def test_to_str(input: Any, expected: str):
     assert to_str(input) == expected
+
+
+@pytest.mark.parametrize(
+    ("dumped_model", "excluded_values", "expected"),
+    [
+        pytest.param(
+            {},
+            (),
+            {},
+        ),
+        pytest.param(
+            {},
+            (None,),
+            {},
+        ),
+        pytest.param(
+            {"foo": 0, "bar": 1},
+            (),
+            {"foo": 0, "bar": 1},
+        ),
+        pytest.param(
+            {"foo": 0, "bar": 1},
+            (None,),
+            {"foo": 0, "bar": 1},
+        ),
+        pytest.param(
+            {"foo": 0, "bar": 1},
+            (0,),
+            {"bar": 1},
+        ),
+        pytest.param(
+            {"foo": 0, "bar": 1},
+            (1,),
+            {"foo": 0},
+        ),
+        pytest.param(
+            {"foo": None},
+            (None,),
+            {},
+        ),
+        pytest.param(
+            {"foo": None, "bar": 0},
+            (None,),
+            {"bar": 0},
+        ),
+        pytest.param(
+            {"foo": None, "bar": 0},
+            (None, 0),
+            {},
+        ),
+    ],
+)
+def test_exclude_by_value(
+    dumped_model: dict[str, Any],
+    excluded_values: tuple[Any, ...],
+    expected: dict[str, Any],
+):
+    assert exclude_by_value(dumped_model, *excluded_values) == expected
+
+
+def test_serialize_as_optional():
+    class Model(SerializeAsOptionalModel):
+        optional_list: SerializeAsOptional[list[str]] = []
+        optional_dict: SerializeAsOptional[dict[str, str]] = {}
+
+    model = Model()
+    assert model.optional_list == []
+    assert model.optional_dict == {}
+    assert model.model_dump() == {"optional_list": None, "optional_dict": None}
+    assert model.model_dump(exclude_defaults=True) == {}
+    assert model.model_dump(exclude_unset=True) == {}
+    # this would fail without inheriting from SerializeAsOptionalModel
+    assert model.model_dump(exclude_none=True) == {}
+
+    model = Model.model_validate({"optional_list": None, "optional_dict": None})
+    assert model.optional_list == []
+    assert model.optional_dict == {}
+
+    model = Model(optional_list=["el"], optional_dict={"foo": "bar"})
+    assert model.optional_list == ["el"]
+    assert model.optional_dict == {"foo": "bar"}
+    assert model.model_dump() == {
+        "optional_list": ["el"],
+        "optional_dict": {"foo": "bar"},
+    }
+    assert model.model_dump(exclude_defaults=True) == {
+        "optional_list": ["el"],
+        "optional_dict": {"foo": "bar"},
+    }
+    assert model.model_dump(exclude_unset=True) == {
+        "optional_list": ["el"],
+        "optional_dict": {"foo": "bar"},
+    }
+    assert model.model_dump(exclude_none=True) == {
+        "optional_list": ["el"],
+        "optional_dict": {"foo": "bar"},
+    }